# Conflicts:
#	pharmacy-service/src/main/java/com/cognizant/pharmacyService/controller/DispenseController.java
#	prescription-service/src/main/java/com/cognizant/prescriptionservice/service/PrescriptionServiceImpl.java

+++ billing-service/src/main/java/com/cognizant/billingService/dto/LabDTO.java
@@ -1,18 +1,18 @@
 package com.cognizant.billingService.dto;
 
-import lombok.Data;
-
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import lombok.Data;
 
 @Data
 public class LabDTO {
 
-    private Long id;
-    private Long patientId;
-    private Long appointmentId;
-    private String testName;
-    private String testCode;
-    private String status;
-    private Double fee;
-    private LocalDateTime createdAt;
+	private Long id;
+	private Long patientId;
+	private Long appointmentId;
+	private String testName;
+	private String testCode;
+	private String status;
+	private BigDecimal fee;
+	private LocalDateTime createdAt;
 }
